Adding test_group to lm-eval configs-7
